remote parameterjsonwriterv2 to fix failing integration tests
