+++ src/Bicep.Core/Emit/ParametersJsonWriter.cs
@@ -3,106 +3,124 @@
 
 using Bicep.Core.Intermediate;
 using Bicep.Core.Semantics;
+using Microsoft.WindowsAzure.ResourceStack.Common.Json;
 using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 
 namespace Bicep.Core.Emit;
 
 public class ParametersJsonWriter
 {
-    private readonly SemanticModel model;
+    private ExpressionBuilder ExpressionBuilder { get; }
 
-    public ParametersJsonWriter(SemanticModel model)
+    private EmitterContext Context => ExpressionBuilder.Context;
+
+    public ParametersJsonWriter(SemanticModel semanticModel)
     {
-        this.model = model;
+        ExpressionBuilder = new ExpressionBuilder(new EmitterContext(semanticModel));
     }
 
-    public void Write(JsonTextWriter jsonWriter)
+    public void Write(SourceAwareJsonTextWriter writer)
     {
-        jsonWriter.WriteStartObject();
-
-        jsonWriter.WritePropertyName("$schema");
-        jsonWriter.WriteValue("https://schema.management.azure.com/schemas/2019-04-01/deploymentParameters.json#");
-
-        jsonWriter.WritePropertyName("contentVersion");
-        jsonWriter.WriteValue("1.0.0.0");
-
-        jsonWriter.WritePropertyName("parameters");
-        jsonWriter.WriteStartObject();
-
-        foreach (var assignment in model.Root.ParameterAssignments)
-        {
-            jsonWriter.WritePropertyName(assignment.Name);
-            jsonWriter.WriteStartObject();
-
-            var parameter = model.EmitLimitationInfo.ParameterAssignments[assignment];
-
-            if (parameter.KeyVaultReferenceExpression is { } keyVaultReference)
-            {
-                WriteKeyVaultReference(jsonWriter, keyVaultReference);
-            }
-            else if (parameter.Value is { } value)
-            {
-                jsonWriter.WritePropertyName("value");
-                value.WriteTo(jsonWriter);
-            }
-            else
-            {
-                throw new InvalidOperationException($"The '{assignment.Name}' parameter assignment defined neither a concrete value nor a key vault reference");
-            }
-
-            jsonWriter.WriteEndObject();
-        }
-
-        jsonWriter.WriteEndObject();
-
-        if (model.Features is { ExtensibilityEnabled: true, ModuleExtensionConfigsEnabled: true })
-        {
-            WriteExtensionConfigs(jsonWriter);
-        }
-
-        jsonWriter.WriteEndObject();
+        var parametersJToken = GenerateParametersJToken(writer.TrackingJsonWriter);
+        parametersJToken.WriteTo(writer);
     }
 
-    private void WriteExtensionConfigs(JsonTextWriter jsonWriter)
+    private JToken GenerateParametersJToken(PositionTrackingJsonTextWriter jsonWriter)
     {
-        jsonWriter.WritePropertyName("extensionConfigs");
+        var emitter = new ExpressionEmitter(jsonWriter, this.Context);
+        
         jsonWriter.WriteStartObject();
-
-        foreach (var extension in model.Root.ExtensionConfigAssignments)
+        emitter.EmitProperty("$schema", "https://schema.management.azure.com/schemas/2019-04-01/deploymentParameters.json#");
+        emitter.EmitProperty("contentVersion", "1.0.0.0");
+        emitter.EmitObjectProperty("parameters", () =>
         {
-            jsonWriter.WritePropertyName(extension.Name);
-            jsonWriter.WriteStartObject();
+            foreach (var assignment in this.Context.SemanticModel.Root.ParameterAssignments)
+            {
+                emitter.EmitObjectProperty(assignment.Name, () => 
+                {
+                    var parameter = this.Context.SemanticModel.EmitLimitationInfo.ParameterAssignments[assignment];
 
-            // TODO(kylealbert): write config properties
+                    if (parameter.KeyVaultReferenceExpression is { } keyVaultReference)
+                    {
+                        WriteKeyVaultReference(emitter, keyVaultReference);
+                    }
+                    else if (parameter.Value is { } value)
+                    {
+                        emitter.EmitProperty("value", () => value.WriteTo(jsonWriter));
+                    }
+                    else if (parameter.Expression is { } expression)
+                    {
+                        emitter.EmitProperty("expression", expression);
+                    }
+                    else
+                    {
+                        throw new InvalidOperationException($"The '{assignment.Name}' parameter assignment defined neither a concrete value nor a key vault reference");
+                    }
+                });
+            }
+        });
+        
+        if (this.Context.SemanticModel.Root.ParameterAssignments.Length > 0 &&
+            this.Context.ExternalInputReferences.ExternalInputIndexMap.Count > 0)
+        {
+            emitter.EmitObjectProperty("externalInputs", () =>
+            {
+                foreach (var reference in this.Context.ExternalInputReferences.ExternalInputIndexMap)
+                {
+                    var functionCall = reference.Key;
+                    var index = reference.Value;
 
-            jsonWriter.WriteEndObject();
+                    var expression = (FunctionCallExpression)ExpressionBuilder.Convert(functionCall);
+                    emitter.EmitObjectProperty(index.ToString(), () =>
+                    {
+                        emitter.EmitProperty("kind", expression.Parameters[0]);
+                        emitter.EmitProperty("options", expression.Parameters[1]);
+                    });
+                }
+            });
+        }
+
+        if (this.Context.SemanticModel.Features is { ExtensibilityEnabled: true, ModuleExtensionConfigsEnabled: true })
+        {
+            WriteExtensionConfigs(emitter, this.Context.SemanticModel.Root.ExtensionConfigAssignments);
         }
 
         jsonWriter.WriteEndObject();
+
+        var content = jsonWriter.ToString();
+        return content.FromJson<JToken>();
     }
 
-    private static void WriteKeyVaultReference(JsonWriter jsonWriter, ParameterKeyVaultReferenceExpression expression)
+    private static void WriteKeyVaultReference(ExpressionEmitter emitter, ParameterKeyVaultReferenceExpression keyVaultReference)
     {
-        jsonWriter.WritePropertyName("reference");
-        jsonWriter.WriteStartObject();
-
-        jsonWriter.WritePropertyName("keyVault");
-        jsonWriter.WriteStartObject();
-
-        jsonWriter.WritePropertyName("id");
-        jsonWriter.WriteValue(expression.KeyVaultId);
-
-        jsonWriter.WriteEndObject();
-
-        jsonWriter.WritePropertyName("secretName");
-        jsonWriter.WriteValue(expression.SecretName);
-
-        if (expression.SecretVersion is string secretVersion)
+        emitter.EmitObjectProperty("reference", () =>
         {
-            jsonWriter.WritePropertyName("secretVersion");
-            jsonWriter.WriteValue(secretVersion);
-        }
+            emitter.EmitObjectProperty("keyVault", () =>
+            {
+                emitter.EmitProperty("id", keyVaultReference.KeyVaultId);
+            });
 
-        jsonWriter.WriteEndObject();
+            emitter.EmitProperty("secretName", keyVaultReference.SecretName);
+
+            if (keyVaultReference.SecretVersion is { } secretVersion)
+            {
+                emitter.EmitProperty("secretVersion", secretVersion);
+            }
+        });
+    }
+
+    private void WriteExtensionConfigs(ExpressionEmitter emitter, IEnumerable<ExtensionConfigAssignmentSymbol> extensionConfigAssignments)
+    {
+        emitter.EmitObjectProperty("extensionConfigs", () =>
+        {
+            foreach (var extension in extensionConfigAssignments)
+            {
+                emitter.EmitObjectProperty(extension.Name, () =>
+                {
+                    // TODO(kylealbert): write config properties
+                });
+            }
+        });
     }
 }
